% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_utils.R
\name{error.predict}
\alias{error.predict}
\title{RMSE and Q2 computation}
\usage{
error.predict(
  y,
  py,
  fpca,
  sd.epsilon = 0,
  scores.epsilon = 0,
  rtx.scores = FALSE
)
}
\arguments{
\item{y}{an array with real two dimensional functional outputs.}

\item{py}{Predictions getting with \code{\link{km_Fpca2d}}.
The two first dimensions correspond to data dimensions,
 which are denoted M and N. The third one is the size of the data set.}

\item{fpca}{an object of class \code{\link{Fpca2d}}.}

\item{sd.epsilon}{a value with \code{y} standard deviation threshold. Q2 values associated to
standard deviations less or equal to \code{sd.epsilon} are NA.
The default is 0. This is to avoid infinite values of Q2.}

\item{scores.epsilon}{a value with  \code{y}  scores standard deviation threshold. Q2 values less or equal to
\code{scores.epsilon} are NA. The default is 0. This is to avoid infinite values of Q2.}

\item{rtx.scores}{a logical value. If TRUE, rmse and Q2 of the predicted scores are returned.
The default is FALSE}
}
\value{
\itemize{
  \item if rtx.scores=FALSE, a list with the following items is returned :
        \itemize{
        \item \code{rmse} : rmse of \code{y} prediction.
        \item \code{Q2} : rmse of \code{y} prediction.
        }
  \item if rtx.scores=TRUE, a list with the following items is returned :
        \itemize{
        \item \code{scores} :
           \itemize{
                 \item \code{rmse} : rmse of \code{y} score prediction.
                  \item \code{Q2} : Q2 of \code{y} score prediction.
            }
            \item \code{y} :the same returned list as for rtx.scores=FALSE
        }
}
}
\description{
RMSE and Q2 computation
}
