% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fpca2d_OrthoNormalBsplines.R
\name{Fpca2d.Bsplines}
\alias{Fpca2d.Bsplines}
\title{Two dimensional functional principal component analysis (FPCA)
by using B-splines basis}
\usage{
Fpca2d.Bsplines(
  x,
  z1,
  z2,
  z1.knots,
  z2.knots,
  norder = 2,
  ortho = "GS",
  expand_knots = FALSE,
  p = 1,
  ncoeff = NULL,
  center = TRUE,
  scale. = FALSE,
  ...
)
}
\arguments{
\item{x}{a three dimensional array, which contains two-dimensional data
(images, maps, etc.). The two first dimensions correspond to data dimensions,
which are denoted M and N. The third one is the size of the data set.}

\item{z1, z2}{two vectors of argument values at which the B-spline basis functions are to be evaluated.}

\item{z1.knots, z2.knots}{The full set of knots used to define the basis functions.}

\item{norder}{Order of the spline fit (degree= order-1). The default is 2.}

\item{ortho}{a character string which specify the orthogonalization method.
If it is "Redd", the \code{\link{orthogonalsplinebasis}} package is used.
If it is "GS", the Gram-Schmidt is used. The default is "GS".}

\item{expand_knots}{a boolean. If it is TRUE, knots are expanded
for appropriate number of knots in bsplines (see \code{\link{expand.knots}}). The default is FALSE.}

\item{p}{a value which fixes the total mean proportion of energy (or mean spatial variance).
The number of coefficients (\code{ncoeff}) used for PCA is calibrated according to its value.
The default is 1. If a value is given in \code{ncoeff}, \code{p} is not used.}

\item{ncoeff}{a value which fixes the number of coefficients used for PCA.
The default is \code{NULL}. If it is \code{NULL}, the number of coefficients is calibrated by using the parameter \code{p}.}

\item{center}{a logical value indicating whether the coefficients should be shifted to be zero centered. The default is \code{TRUE}.}

\item{scale.}{a logical value indicating whether the coefficients should be scaled to have unit variance before the analysis takes place.
The default is \code{FALSE}.}

\item{...}{arguments of \code{\link{prcomp}}, which can fix characteristics of PCA.}
}
\value{
an object of class "Fpca2d" which is a list with the following items :
 \itemize{
   \item \code{sdev} : the standard deviations of the principal components
                       (i.e., the square roots of the eigenvalues of the covariance operator).
   \item \code{EigFct} : a three dimensional array with the eigenfunctions.
                         The two first dimensions correspond to data (images, maps, etc.) dimensions.
                         The third one is the maximal number of principal component used.
   \item \code{x} : if \code{retx} is \code{TRUE} (see \code{\link{prcomp}}).
                    Scores (coordinates in the eigen basis) of the data are returned.
   \item \code{center}, \code{scale} : a logical value which indicates if data coefficients are respectively centered or scaled.
 }
}
\description{
Two dimensional functional principal component analysis (FPCA)
by using B-splines basis
}
\seealso{
\code{\link{Fpca2d}}
}
