% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fpca2d_Wavelets.R
\name{Fpca2d.Wavelets}
\alias{Fpca2d.Wavelets}
\title{Two dimensional functional principal component analysis (FPCA)
       by using wavelet basis}
\usage{
Fpca2d.Wavelets(
  x,
  wf,
  J,
  boundary = "periodic",
  p = 1,
  ncoeff = NULL,
  center = TRUE,
  scale. = FALSE,
  ...
)
}
\arguments{
\item{x}{a three dimensional array, which contains two-dimensional data (images, maps, etc.).
The two first dimensions correspond to data dimensions, which are denoted M and N. The third one is the size of the data set.}

\item{wf}{a character string which specifies the wavelet filter (see \code{\link{dwt.2d}}).}

\item{J}{depth of the wavelet decomposition, must be a number less than or equal to log(min(M,N),2).}

\item{boundary}{a character string which specifies how boundaries are treated. Only "periodic" is currently implemented (see \code{\link{dwt.2d}}).}

\item{p}{a value which fixes the total mean proportion of energy (or mean spatial variance).
The number of coefficients (\code{ncoeff}) used for PCA is calibrated according to its value.
The default is 1. If a value is given in \code{ncoeff}, \code{p} is not used.}

\item{ncoeff}{a value which fixes the number of coefficients used for PCA.
The default is \code{NULL}. If it is \code{NULL}, the number of coefficients is calibrated by using the parameter \code{p}.}

\item{center}{a logical value indicating whether the coefficients should be shifted to be zero centered. The default is \code{TRUE}.}

\item{scale.}{a logical value indicating whether the coefficients should be scaled to have unit variance before the analysis takes place.
The default is \code{FALSE}.}

\item{...}{arguments of \code{\link{prcomp}}, which can fix characteristics of PCA.}
}
\value{
an object of class "Fpca2d" which is a list with the following items :
 \itemize{
   \item \code{sdev} : the standard deviations of the principal components
                       (i.e., the square roots of the eigenvalues of the covariance operator).
   \item \code{EigFct} : a three dimensional array with the eigenfunctions.
                         The two first dimensions correspond to data (images, maps, etc.) dimensions.
                         The third one is the maximal number of principal component used.
   \item \code{x} : if \code{retx} is \code{TRUE} (see \code{\link{prcomp}}).
                    Scores (coordinates in the eigen basis) of the data are returned.
   \item \code{center}, \code{scale} : a logical value which indicates if data coefficients are respectively centered or scaled.
 }
}
\description{
performs FPCA on two-dimensional functional data (data set of images, maps, ...).
             The implementation of FPCA is based on wavelet basis.
}
\seealso{
\code{\link{Fpca2d}}
}
