% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrthoNormalBsplines2D_utils.R
\name{Inverse2D.OrthoNormalBsplines2D}
\alias{Inverse2D.OrthoNormalBsplines2D}
\title{Inverse transform of control points of two-dimensional B-splines basis}
\usage{
\method{Inverse2D}{OrthoNormalBsplines2D}(object, coeff, ...)
}
\arguments{
\item{object}{an object of class \code{\link{OrthoNormalBsplines2D}}
with functions of the B-splines basis.}

\item{coeff}{matrix with control points on two-dimensional B-splines basis.}

\item{...}{other arguments.}
}
\value{
an array with two-dimensional approximated data.
}
\description{
Inverse transform of control points of two-dimensional B-splines basis
}
\seealso{
\code{\link{coef.OrthoNormalBsplines2D}} \code{\link{OrthoNormalBsplines2D}}
\code{\link{Inverse2D}}
}
