% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Inverse2D}
\alias{Inverse2D}
\title{Inverse transform of wavelet or orthonormal B-splines basis.}
\usage{
Inverse2D(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{Wavelet2D}} or \code{\link{OrthoNormalBsplines2D}}.}

\item{...}{if \code{object} is an object of class \code{OrthoNormalBsplines2D},
the matrix of control points must be given (see \code{\link{Inverse2D.OrthoNormalBsplines2D}}).}
}
\value{
a three dimensional array.
The first two dimensions correspond to maps dimensions.
The third dimension corresponds to the size of the data set.
}
\description{
embeds the coefficients of "Wavelet2D" or "OrthonormalBsplines2D" basis
onto two-dimensional domain.
}
\seealso{
\code{\link{Inverse2D.OrthoNormalBsplines2D}} \code{\link{Inverse2D.Wavelet2D}}
}
\author{
Tran Vi-vi Elodie PERRIN
}
