% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fpca2d_utils.R
\name{inverse_Fpca2d}
\alias{inverse_Fpca2d}
\title{Get approximation with scores of Fpca2d}
\usage{
inverse_Fpca2d(scores, fpca)
}
\arguments{
\item{scores}{a matrix where columns correpond to scores of principal components.}

\item{fpca}{an object of class \code{Fpca2d}.}
}
\value{
approximation of scores.
}
\description{
Get approximation with scores of Fpca2d
}
\keyword{internal}
