% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanPoe.R
\name{MeanPoe.coef_OrthoNormalBsplines2D}
\alias{MeanPoe.coef_OrthoNormalBsplines2D}
\title{Mean proportion of energy of B-splines control points}
\usage{
\method{MeanPoe}{coef_OrthoNormalBsplines2D}(object)
}
\arguments{
\item{object}{control point outputs of  \code{\link{coef.OrthoNormalBsplines2D}} .}
}
\value{
a matrix K×L which contains the mean proportion of energy of each coefficient,
        with K×L is the dimensions of the B-splines basis
}
\description{
Mean proportion of energy of B-splines control points
}
\keyword{internal}
