% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MeanPoe}
\alias{MeanPoe}
\title{Mean proportion of energy}
\usage{
MeanPoe(object)
}
\arguments{
\item{object}{an object of class \code{\link{Wavelet2D}} or \code{CoefOrthonormalBsplines2D}
which correspond to coefficients from wavelets or Bsplines basis.}
}
\value{
a matrix K×L which contains the mean proportion of energy of each coefficient,
        with K×L is the dimensions of the functional basis.
}
\description{
Wrapper function to compute mean proportion of energy for given coefficents.
The valid objects are "Wavelet2D" and "CoefOrthonormalBsplines2D".
}
\author{
Tran Vi-vi Elodie PERRIN
}
\keyword{internal}
