% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanPoe.R
\name{MeanPoe.Wavelet2D}
\alias{MeanPoe.Wavelet2D}
\title{Mean proportion of energy of wavelet coefficients}
\usage{
\method{MeanPoe}{Wavelet2D}(object)
}
\arguments{
\item{object}{an object of class \code{\link{Wavelet2D}}.}
}
\value{
a matrix K×L which contains the mean proportion of energy of each coefficient,
        with K×L is the dimensions of the wavelet basis
}
\description{
Mean proportion of energy of wavelet coefficients
}
\keyword{internal}
