% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gp_Fpca2d}
\alias{predict.gp_Fpca2d}
\title{Predict values and confidence intervals at newdata for a km_Fpca2d object}
\usage{
\method{predict}{gp_Fpca2d}(object, newdata, type, compute = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{\link{gp_Fpca2d}}.}

\item{newdata}{a vector, matrix or data frame containing the points where to perform predictions.}

\item{type}{a character string corresponding to the kriging family, to be chosen between simple kriging ("SK"), or universal kriging ("UK").}

\item{compute}{an optional boolean. If FALSE, only the kriging mean is computed. If TRUE, the kriging variance (here, the standard deviation is returned) and confidence intervals are computed too.}

\item{...}{see \code{\link{predict.km}}.}
}
\value{
a list with the following items :
\itemize{
  \item \code{scores_predict} :  prediction of scores from two-dimensional FPCA
  (see  \code{\link{Fpca2d}} and \code{\link{predict.km}}).
  \item \code{mean} : an array with three dimensions which contains image (or map) kriging means.
  The two first dimensions corresponds to the output data dimensions. The third correspond to the number of predictions.
 \item \code{sd} : it is return only if compute = TRUE.
 If TRUE, an array which contains image (or map) prediction standard deviation.
 \item \code{lower95,upper95} :  bounds of the 95% confidence interval computed at \code{newdata}.
  Not computed if compute=FALSE.
}
}
\description{
Predict values and confidence intervals at newdata for a km_Fpca2d object
}
\seealso{
\code{\link{predict.km}}
}
